YANP-12 Change Account Id to an Int. It's Int(11) in the database which is 4 bytes max

+++ src/main/groovy/com/woodlands/yanp/auth/db/entity/AccountEntity.groovy
@@ -35,7 +35,7 @@ class AccountEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = 'ID', nullable = false)
-    Long id
+    Integer id
 
     @JsonProperty
     @Column(name = 'USERNAME', length = 32, unique = true)
